Modified parsed_version to keep as string instead of float
The current parsed_version line does not work on versions with trailing zeros... etc 4.10. When converting to float, the trailing zero is dropped and the version becomes 4.1, which is too low by definition.

This commit modified the parsing to parse out numbers from the string but keep them as a string. Now the version numbers are evaludated as strings (i.e. '4.10').