Refactor init and config
Make the __init__.py and config more concise with appropriate Flask
variables. There was a lot of duplication of code from the config object
being read and in the initialization under `create_app`. Also removed
heroku-related code in these two files as app is no longer hosted on
heroku.

Additionally updated the .env.template to reflect changes from the code.