Update dev instructions
Old instructions for testing development was outdated, referring to a
now obsolete `manage.py` file. This commit updates this, referring the
user to make use of docker compose for testing, which is the easiest
way, albeit at the cost of additional installations.

The benefit of this is the user does not need to separately install and
setup Postgres. However, this also means to test each change, the user
will have to run a separate `docker compose up --build`.